dashboard

+++ backend_rails_api/db/migrate/001_create_licenses.rb
@@ -2,7 +2,7 @@ class CreateLicenses < ActiveRecord::Migration[6.0]
   def change
     create_table :licenses do |t|
       t.string :name
-      t.integer :cost
+      t.float :cost
 
       t.timestamps
     end
